requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
torch
torchvision
networkx
matplotlib
requests
numpy
pillow
